git-svn-id: https://sourceforge.llnl.gov/svn/repos/samrai_project/trunk/MASTER@858 f45145c3-882e-4954-aa44-00140844b9ef

+++ source/SAMRAI/xfer/RefineScheduleConnectorWidthRequestor.C
@@ -145,7 +145,7 @@ RefineScheduleConnectorWidthRequestor::computeRequiredFineConnectorWidthsForRecu
    const hier::PatchHierarchy& patch_hierarchy,
    int initial_dst_ln) const
 {
-  if ( (int)fine_connector_widths.size() < initial_dst_ln ) {
+   if ( fine_connector_widths.size() < initial_dst_ln ) {
       fine_connector_widths.insert(
          fine_connector_widths.end(),
          initial_dst_ln - fine_connector_widths.size(),
